Move ServerResponse class inside http namespace
