+++ src/Entities/Payments/PreCheckoutQuery.php
@@ -48,7 +48,7 @@ class PreCheckoutQuery extends Entity
      * @param bool $ok
      * @param array $data
      *
-     * @return \Longman\TelegramBot\Entities\ServerResponse
+     * @return \Longman\TelegramBot\Http\ServerResponse
      */
     public function answer($ok, array $data = [])
     {
